Update Session.java
Added methods to accept and return key handle as parameters. It was observed that with most of the HSM's if we want to perform multiple key operations within a single session like derive and wrap, doing a findObject every time seems to very costly. Instead we can do findObject once within a given session and use the handle elsewhere of the returned object. This seems to have around 3 times better performance.